Migrated the HttpContentDeserializerExtensions tests.
